+++ govcd/api_test.go
@@ -31,12 +31,14 @@ At least one of the following tags should be defined:
    * catalog:    Runs catalog related tests (also catalog_item, media)
    * disk:       Runs disk related tests
    * extnetwork: Runs external network related tests
+   * lb:       	 Runs load balancer related tests
    * network:    Runs network related tests
    * gateway:    Runs edge gateway related tests
    * org:        Runs org related tests
    * query:      Runs query related tests
    * system:     Runs system related tests
    * task:       Runs task related tests
+   * user:       Runs user related tests
    * vapp:       Runs vapp related tests
    * vdc:        Runs vdc related tests
    * vm:         Runs vm related tests